+++ app/models/user.rb
@@ -3,9 +3,9 @@
 # Table name: users
 #
 #  id              :integer          not null, primary key
-#  email_address   :string(255)
-#  name            :string(255)
-#  password_digest :string(255)
+#  email_address   :string
+#  name            :string
+#  password_digest :string
 #  created_at      :datetime         not null
 #  updated_at      :datetime         not null
 #